Move helpers fixtures in to `mavis.test`
This moves the fixtures related to the various helpers in to their
own file.
