+++ mavis/test/__init__.py
@@ -1,6 +1,3 @@
-import pytest
-
-from .data import TestData
 from .hooks import pytest_runtest_logreport, pytest_sessionfinish, pytest_sessionstart
 from .fixtures import (
     base_url,
@@ -9,7 +6,10 @@ from .fixtures import (
     children_page,
     consent_page,
     dashboard_page,
+    get_online_consent_url,
     import_records_page,
+    log_in_as_admin,
+    log_in_as_nurse,
     log_in_page,
     playwright_operations,
     programmes_page,
@@ -17,6 +17,7 @@ from .fixtures import (
     screenshots_path,
     sessions_page,
     start_page,
+    test_data,
     unmatched_page,
     vaccines_page,
 )
@@ -44,7 +45,10 @@ __all__ = [
     "clinics",
     "consent_page",
     "dashboard_page",
+    "get_online_consent_url",
     "import_records_page",
+    "log_in_as_admin",
+    "log_in_as_nurse",
     "log_in_page",
     "nurse",
     "onboard",
@@ -63,12 +67,8 @@ __all__ = [
     "start_page",
     "superuser",
     "team",
+    "test_data",
     "unmatched_page",
     "users",
     "vaccines_page",
 ]
-
-
-@pytest.fixture(scope="session")
-def test_data(organisation, schools):
-    return TestData(organisation, schools)
